+++ atm-e2e-tests/test/pageobjects/deposit/confirm.page.js
@@ -1,7 +1,7 @@
 import BasePage, { tag } from '../base.page.js';
 
-class TopupConfirmPage extends BasePage {
-    get screen()     { return tag('screen_topupConfirm') }
+class DepositConfirmPage extends BasePage {
+    get screen()     { return tag('screen_confirmVerifyMoney') }
     get confirmBtn() { return tag('btn_ยืนยัน') }
     get cancelBtn()  { return tag('btn_ยกเลิก') }
 
@@ -11,4 +11,4 @@ class TopupConfirmPage extends BasePage {
     }
 }
 
-export default new TopupConfirmPage();
+export default new DepositConfirmPage();